refactor: remove direct file deletion from `ArtifactStorageService`

+++ src/services/artifact.storage.service.ts
@@ -4,7 +4,7 @@ import * as ProjectDAL from "@dal/project.dal";
 import * as ReleaseDAL from "@dal/release.dal";
 import { ArtifactDTO } from "@models/release.model";
 import { ReleaseArtifactIdPathParams, ReleaseIdPathParam } from "@schemas/release.schema";
-import { FileStorageOperation } from "@utils/common";
+import { StorageManager } from "@services/storage.manager";
 import { ApiError } from "@utils/error";
 import { Logger } from "@utils/logger";
 import * as Utils from "@utils/utils";
@@ -17,12 +17,12 @@ import { ReleaseCacheService } from "./cache.service";
 
 const storage = multer.diskStorage({
   destination: function (_, __, callback) {
-    callback(null, FileStorageOperation.DEFAULT_FILE_STORAGE_PATH);
+    callback(null, StorageManager.DEFAULT_STORAGE_DIRECTORY_PATH);
   },
   filename: function (_, file, callback) {
     const id = crypto.randomUUID();
-    const originalFilename = file.originalname;
-    const uniqueFilename = `$${id}-${originalFilename}`;
+    const timestamps = Date.now();
+    const uniqueFilename = `${timestamps}-${id}${path.extname(file.originalname)}`;
     callback(null, uniqueFilename);
   },
 });
@@ -30,10 +30,6 @@ const storage = multer.diskStorage({
 export const MAX_FILE_SIZE_IN_BYTES = 8e6; // 8mb in bytes....
 export const upload = multer({ storage: storage });
 
-/**
- * TODO: [PRI-02] Remove the FileStorageOperation operations from this service....
- */
-// ** [DONE] implement file dedup checks.......
 @injectable()
 export class ArtifactStorageService {
   constructor(
@@ -41,7 +37,7 @@ export class ArtifactStorageService {
     private readonly _cache: ReleaseCacheService,
   ) {}
 
-  public async deleteBlob(payload: ReleaseArtifactIdPathParams) {
+  public async remove(payload: ReleaseArtifactIdPathParams): Promise<void> {
     const { releaseId, artifactId } = payload;
     const release = await ReleaseDAL.findReleaseByPublicId(releaseId);
     if (!release) {
@@ -51,22 +47,20 @@ export class ArtifactStorageService {
     const artifact = await ArtifactDAL.findArtifactById(artifactId);
 
     if (!artifact) {
-      throw new ApiError("File not found!", StatusCodes.NOT_FOUND);
+      throw new ApiError("Artifact not found!", StatusCodes.NOT_FOUND);
     }
 
-    if (release.isRevoked()) {
-      throw new ApiError("Revoked release artifact cannot be deleted!", StatusCodes.CONFLICT);
-    }
-
-    if (release.isProduction()) {
-      throw new ApiError("Production release artifact cannot be deleted!", StatusCodes.CONFLICT);
+    if (release.isRevoked() || release.isProduction()) {
+      throw new ApiError(
+        `${release.getChannel()} release artifact cannot be deleted!`,
+        StatusCodes.CONFLICT,
+      );
     }
 
     const transaction = await db.transaction();
     try {
       await ArtifactDAL.deleteByFilename(artifact.getFileName(), transaction);
       await ReleaseDAL.updateReleaseChannel(releaseId, "draft", transaction);
-      await FileStorageOperation.removePermanently(artifact.getFileName(), "storage");
 
       const project = await ProjectDAL.findProjectByInternalId(release.getProjectForeignKeyId());
 
@@ -76,33 +70,21 @@ export class ArtifactStorageService {
         project!.getPublicId(),
         "staging",
       );
+
       if (isInvalidated) {
         Logger.info(`Invalidated artifact ${cacheKey} from cache!`);
       }
+
       await transaction.commit();
       Logger.info("Artifact deleted!");
     } catch (error) {
       Logger.error("Artifact Deletion Error ", error as string);
       await transaction.rollback();
+      throw error;
     }
   }
 
-  public async findUploadedArtifacts(releaseId: string) {
-    const release = await ReleaseDAL.findReleaseByPublicId(releaseId);
-    if (!release) {
-      throw new ApiError("Release not found!", StatusCodes.BAD_REQUEST);
-    }
-
-    const artifact = await ArtifactDAL.findArtifactByInternalId(release.getId());
-
-    if (!artifact) {
-      throw new ApiError("artifact not found!", StatusCodes.NOT_FOUND);
-    }
-
-    return artifact;
-  }
-
-  public async getBinaryFilePath(params: ReleaseArtifactIdPathParams) {
+  public async getBinaryFilePath(params: ReleaseArtifactIdPathParams): Promise<string> {
     const { releaseId, artifactId } = params;
     const release = await ReleaseDAL.findReleaseByPublicId(releaseId);
     if (!release) {
@@ -116,7 +98,7 @@ export class ArtifactStorageService {
     }
 
     const pathToFile = path.join(
-      FileStorageOperation.DEFAULT_FILE_STORAGE_PATH,
+      StorageManager.DEFAULT_STORAGE_DIRECTORY_PATH,
       artifact.getFileName(),
     );
 
@@ -136,18 +118,21 @@ export class ArtifactStorageService {
       throw new ApiError("Release not found!", StatusCodes.BAD_REQUEST);
     }
 
-    /**
-     * * If new file is uploaded and has no conflict with prior release artifacts then it gets saved to db....
-     * * If again same file is uploaded then "Same artifact already exists!" error is thrown...
-     * * If new file is uploaded after first successful save then "Another artifact already exists!" error is thrown..
-     */
-    const hash = await Utils.generateFileHash(path.join("storage", file.filename));
-
-    if (await ArtifactDAL.isHashFound(release.getId(), hash)) {
-      const artifact = await ArtifactDAL.findPendingArtifact(release.getId(), file.originalname);
-      await artifact?.destroy({ force: true });
-      await this.deleteBinaryFromStorage(artifact!.filename!);
+    if (release.isProduction() || release.isStaging()) {
+      throw new ApiError(
+        `Cannot upload artifact on a ${release.getChannel()} mode release!`,
+        StatusCodes.CONFLICT,
+      );
     }
+
+    // *on new file upload remove the old artifact record....
+    const artifactData = await ArtifactDAL.findArtifactByReleaseId(release.getId());
+    if (artifactData) {
+      await artifactData.destroy({ force: true });
+    }
+
+    const pathToFile = path.join("storage", file.filename);
+    const hash = await Utils.generateFileHash(pathToFile);
 
     const collisionRelease = await ProjectDAL.isHashFoundAcrossReleases(
       release.getProjectForeignKeyId(),
@@ -155,7 +140,6 @@ export class ArtifactStorageService {
     );
 
     if (collisionRelease) {
-      await this.deleteBinaryFromStorage(file.filename);
       throw new ApiError(
         `Collision detected with release v${collisionRelease.Releases![0].version}!`,
         StatusCodes.CONFLICT,
@@ -170,15 +154,5 @@ export class ArtifactStorageService {
     });
 
     return artifact.toDTO();
-  }
-
-  public async deleteBinaryFromStorage(filename: string): Promise<void> {
-    try {
-      await FileStorageOperation.removePermanently(filename, "storage");
-      Logger.info("file deleted! ".concat(filename));
-    } catch (error) {
-      Logger.error((error as Error).message);
-      throw error;
-    }
   }
 }
